add some entitity

+++ src/main/java/com/example/estoutic/controller/models/TelephoneNumberSaveRequest.java
@@ -3,9 +3,11 @@ package com.example.estoutic.controller.models;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Data;
 
+import java.math.BigInteger;
+
 @Data
 public class TelephoneNumberSaveRequest {
     @JsonProperty("telephone")
-    private Long telephone;
+    private BigInteger telephone;
 
 }
